requirements.txt
azure-ai-textanalytics==5.2.0b2
azure-cognitiveservices-speech==1.21.0
azure-common==1.1.28
azure-core==1.23.1
azure-search-documents==11.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
idna==3.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.6.21
oauthlib==3.2.0
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
